InvalidParameterInfo class refactoring
